+++ apps/backend/src/modules/friends/friend.repository.ts
@@ -1,5 +1,6 @@
 import { DatabaseTableName } from "~/libs/modules/database/database.js";
 import { type Repository } from "~/libs/types/types.js";
+import { EMPTY_ARRAY_LENGTH } from "~/libs/types/types.js";
 import { UserEntity } from "~/modules/users/user.entity.js";
 import { type UserModel } from "~/modules/users/user.model.js";
 
@@ -52,6 +53,7 @@ class FriendRepository implements Repository<UserEntity> {
 			firstName: followingUser.userDetails.firstName,
 			id: followingUser.id,
 			lastName: followingUser.userDetails.lastName,
+			nickname: followingUser.userDetails.nickname,
 			passwordHash: followingUser.passwordHash,
 			passwordSalt: followingUser.passwordSalt,
 			updatedAt: followingUser.updatedAt,
@@ -105,6 +107,7 @@ class FriendRepository implements Repository<UserEntity> {
 					firstName: user.userDetails.firstName,
 					id: user.id,
 					lastName: user.userDetails.lastName,
+					nickname: user.userDetails.nickname,
 					passwordHash: user.passwordHash,
 					passwordSalt: user.passwordSalt,
 					updatedAt: user.updatedAt,
@@ -131,11 +134,30 @@ class FriendRepository implements Repository<UserEntity> {
 				firstName: user.userDetails.firstName,
 				id: user.id,
 				lastName: user.userDetails.lastName,
+				nickname: user.userDetails.nickname,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
 				updatedAt: user.updatedAt,
 			});
 		});
+	}
+
+	public async getIsFollowing(
+		currentUserId: number,
+		otherUserId: number,
+	): Promise<boolean> {
+		const userFollowings = await this.userModel
+			.query()
+			.leftJoin(
+				DatabaseTableName.FRIENDS,
+				`${DatabaseTableName.USERS}.id`,
+				"=",
+				`${DatabaseTableName.FRIENDS}.following_id`,
+			)
+			.where(`${DatabaseTableName.FRIENDS}.follower_id`, "=", currentUserId)
+			.where(`${DatabaseTableName.FRIENDS}.following_id`, "=", otherUserId);
+
+		return userFollowings.length > EMPTY_ARRAY_LENGTH;
 	}
 
 	public async getIsSubscribedByRequestId(
@@ -190,6 +212,7 @@ class FriendRepository implements Repository<UserEntity> {
 				firstName: user.userDetails.firstName,
 				id: user.id,
 				lastName: user.userDetails.lastName,
+				nickname: user.userDetails.nickname,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
 				updatedAt: user.updatedAt,
@@ -217,6 +240,7 @@ class FriendRepository implements Repository<UserEntity> {
 				firstName: user.userDetails.firstName,
 				id: user.id,
 				lastName: user.userDetails.lastName,
+				nickname: user.userDetails.nickname,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
 				updatedAt: user.updatedAt,
@@ -244,6 +268,7 @@ class FriendRepository implements Repository<UserEntity> {
 				firstName: user.userDetails.firstName,
 				id: user.id,
 				lastName: user.userDetails.lastName,
+				nickname: user.userDetails.nickname,
 				passwordHash: user.passwordHash,
 				passwordSalt: user.passwordSalt,
 				updatedAt: user.updatedAt,
@@ -266,6 +291,7 @@ class FriendRepository implements Repository<UserEntity> {
 			firstName: updatedSubscription.userDetails.firstName,
 			id: updatedSubscription.id,
 			lastName: updatedSubscription.userDetails.lastName,
+			nickname: updatedSubscription.userDetails.nickname,
 			passwordHash: updatedSubscription.passwordHash,
 			passwordSalt: updatedSubscription.passwordSalt,
 			updatedAt: updatedSubscription.updatedAt,