+++ apps/backend/src/modules/users/libs/types/types.ts
@@ -1,6 +1,7 @@
 export {
 	type UserAuthResponseDto,
 	type UserGetAllResponseDto,
+	type UserGetByIdRequestDto,
 	type UserProfileRequestDto,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,