+++ packages/shared/src/modules/users/libs/validation-schemas/user-profile.validation-schema.ts
@@ -5,6 +5,12 @@ import { UserValidationMessage, UserValidationRule } from "../enums/enums.js";
 type UserProfileRequestValidationDto = {
 	firstName: z.ZodString;
 	lastName: z.ZodString;
+	nickname: z.ZodNullable<
+		z.ZodPipeline<
+			z.ZodEffects<z.ZodString, null | string, string>,
+			z.ZodUnion<[z.ZodString, z.ZodLiteral<null>]>
+		>
+	>;
 };
 
 const userProfile = z.object<UserProfileRequestValidationDto>({
@@ -62,6 +68,33 @@ const userProfile = z.object<UserProfileRequestValidationDto>({
 		.max(UserValidationRule.LAST_NAME_MAXIMUM_LENGTH, {
 			message: UserValidationMessage.LAST_NAME_MAXIMUM_LENGTH,
 		}),
+	nickname: z
+		.string()
+		.transform((value) => {
+			return value === "" ? null : value;
+		})
+		.pipe(
+			z
+				.string()
+				.trim()
+				.min(UserValidationRule.FIELD_MINIMUM_LENGTH, {
+					message: UserValidationMessage.FIELD_REQUIRE,
+				})
+				.regex(/^[\d_a-z]+$/, {
+					message: UserValidationMessage.NICKNAME_INVALID_CHARACTERS,
+				})
+				.regex(/[a-z]/, {
+					message: UserValidationMessage.NICKNAME_AT_LEAST_ONE_LETTER,
+				})
+				.min(UserValidationRule.NICKNAME_MINIMUM_LENGTH, {
+					message: UserValidationMessage.NICKNAME_MINIMUM_LENGTH,
+				})
+				.max(UserValidationRule.NICKNAME_MAXIMUM_LENGTH, {
+					message: UserValidationMessage.NICKNAME_MAXIMUM_LENGTH,
+				})
+				.or(z.literal(null)),
+		)
+		.nullable(),
 });
 
 export { userProfile };